fix: append mode while getting file content doesn't allow reading in write protected environments
This fixes the issue of not being able to read the config file when it's
only write restricted. I encountered this when handling the config file
using nix.

I have checked all the functionalities with ipc and they all work fine.
No need to open in append mode if nothing is being written to the config
interactively.